test(ItineraryExistence): Add test override to default OTP response provider.
